Added Circle.js - this file will handle everything to do with Circles, drawing them and all the fancy maths for circles simplified.
Small update to Korganic.js, it now loads the Canvas and Context. 

Update to index.html it now loads Korganic, nothing special, hopefully in the next coming months I will be adding the GUI interface for Korganic, but for now the index is just a testing ground for features added as I plod along.